requirements.txt
# mdaudiobook - Core Dependencies
# Professional Markdown to Audiobook Pipeline

# Core Processing
markdown>=3.4.4
pyyaml>=6.0
python-frontmatter>=1.0.0
jinja2>=3.1.2

# Text Processing and Enhancement
regex>=2023.6.3
nltk>=3.8.1
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Audio Processing
pydub>=0.25.1
mutagen>=1.47.0
librosa>=0.10.1
soundfile>=0.12.1

# TTS Providers (Python packages only)
# Note: piper-tts, espeak-ng, festival are system packages
# Install via: sudo apt install espeak-ng festival piper-tts (or equivalent).

# TTS Providers - API (optional)
elevenlabs>=0.2.24
azure-cognitiveservices-speech>=1.31.0
google-cloud-texttospeech>=2.14.1

# AI Enhancement - Local
ollama>=0.1.7
transformers>=4.33.2
torch>=2.0.1

# AI Enhancement - API (optional)
openai>=0.28.0
anthropic>=0.3.11

# HTTP and API
requests>=2.31.0
httpx>=0.24.1
aiohttp>=3.8.5

# Utilities
click>=8.1.7
rich>=13.5.2
tqdm>=4.66.1
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Configuration and Validation
pydantic>=2.3.0
jsonschema>=4.19.0
cerberus>=1.3.5

# Logging and Monitoring
loguru>=0.7.0
structlog>=23.1.0

# Development and Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.3

# Documentation
mkdocs>=1.5.2
mkdocs-material>=9.2.3

# Platform-specific dependencies
# Windows
pywin32>=306; sys_platform == "win32"

# macOS
pyobjc-framework-AVFoundation>=9.2; sys_platform == "darwin"

# Linux audio
# Note: python3-alsaaudio should be installed via system package manager:
# sudo apt install python3-alsaaudio (Ubuntu/Debian)
# sudo yum install python3-alsaaudio (RHEL/CentOS)
# python-alsaaudio>=0.9.2; sys_platform == "linux"  # Not available via pip
